backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-filter==25.2
django-cors-headers==4.9.0
psycopg[binary]==3.2.12
